+++ fml/shared_thread_merger_impl.h
@@ -36,12 +36,28 @@ class SharedThreadMergerImpl
   static fml::RefPtr<SharedThreadMergerImpl> GetSharedImpl(
       fml::TaskQueueId owner,
       fml::TaskQueueId subsumed);
+  // It's called by |RasterThreadMerger::RecordMergerCaller()|.
+  // See the doc of |RasterThreadMerger::RecordMergerCaller()|.
   void RecordMergerCaller(RasterThreadMerger* caller);
-  bool MergeWithLease(size_t lease_term);
-  /// TODO DOC
+
+  // It's called by |RasterThreadMerger::MergeWithLease()|.
+  // See the doc of |RasterThreadMerger::MergeWithLease()|.
+  bool MergeWithLease(int lease_term);
+
+  // It's called by |RasterThreadMerger::UnMergeNowIfLastOne()|.
+  // See the doc of |RasterThreadMerger::UnMergeNowIfLastOne()|.
   bool UnMergeNowIfLastOne(RasterThreadMerger* caller);
-  void ExtendLeaseTo(size_t lease_term);
+
+  // It's called by |RasterThreadMerger::ExtendLeaseTo()|.
+  // See the doc of |RasterThreadMerger::ExtendLeaseTo()|.
+  void ExtendLeaseTo(int lease_term);
+
+  // It's called by |RasterThreadMerger::IsMergedUnSafe()|.
+  // See the doc of |RasterThreadMerger::IsMergedUnSafe()|.
   bool IsMergedUnSafe() const;
+
+  // It's called by |RasterThreadMerger::DecrementLease()|.
+  // See the doc of |RasterThreadMerger::DecrementLease()|.
   bool DecrementLease();
 
  private:
